+++ api/lib/reducers/polkadotV0-reducers.js
@@ -7,6 +7,7 @@ const {
 } = require('../../common/numbers.js')
 const { getProposalSummary } = require('./common')
 const { lunieMessageTypes } = require('../../lib/message-types')
+const { hexToString } = require('@polkadot/util')
 
 const CHAIN_TO_VIEW_COMMISSION_CONVERSION_FACTOR = 1e-9
 
@@ -223,6 +224,8 @@ function getMessageType(section, method) {
   switch (`${section}.${method}`) {
     case 'balances.transfer':
       return lunieMessageTypes.SEND
+    case 'balances.transferKeepAlive':
+      return lunieMessageTypes.SEND
     case 'staking.payoutStakers':
       return lunieMessageTypes.CLAIM_REWARDS
     case 'lunie.staking':
@@ -281,22 +284,30 @@ async function parsePolkadotTransaction(
   }
 }
 
-function getExtrinsicSuccess(extrinsicIndex, blockEvents, isBatch) {
-  const events = blockEvents.filter(({ phase }) => {
-    return parseInt(phase.toHuman().ApplyExtrinsic) === extrinsicIndex // index is a string
-  })
+function transactionReducerV2(network, extrinsic, blockHeight, reducers) {
+  const hash = extrinsic.hash
+  const signer = extrinsic.signature === null ? '' : extrinsic.signature.signer
+  const isBatch =
+    extrinsic.method.pallet === `utility` && extrinsic.method.method === `batch`
+  const messages = aggregateLunieStaking(
+    isBatch ? extrinsic.args.calls : [extrinsic]
+  )
+
   // if tx is a batch, we need to check if all of the batched txs went through
+  let success
   if (isBatch) {
-    return !!events.find(
-      ({ event }) =>
-        event.section === `utility` && event.method === `BatchCompleted`
+    success = !!extrinsic.events.find(
+      (event) =>
+        event.method.pallet === `utility` &&
+        event.method.method === `BatchCompleted`
+    )
+  } else {
+    success = !!extrinsic.events.find(
+      (event) =>
+        event.method.pallet === `system` &&
+        event.method.method === `ExtrinsicSuccess`
     )
   }
-  return !!events.find(
-    ({ event }) =>
-      event.section === `system` && event.method === `ExtrinsicSuccess`
-  )
-}
 
 async function transactionReducerV2(
   network,
@@ -352,8 +363,8 @@ function aggregateLunieStaking(messages) {
   let reducedMessages = []
   messages.forEach((current) => {
     if (
-      current.toHuman().section === 'staking' &&
-      current.toHuman().method === 'bond'
+      current.method.pallet === 'staking' &&
+      current.method.method === 'bond'
     ) {
       aggregatedLunieStaking.amount =
         aggregatedLunieStaking.amount + current.args.value
@@ -361,8 +372,8 @@ function aggregateLunieStaking(messages) {
     }
 
     if (
-      current.toHuman().section === 'staking' &&
-      current.toHuman().method === 'bondExtra'
+      current.method.pallet === 'staking' &&
+      current.method.method === 'bondExtra'
     ) {
       aggregatedLunieStaking.amount =
         aggregatedLunieStaking.amount + current.args.max_additional
@@ -370,18 +381,18 @@ function aggregateLunieStaking(messages) {
     }
 
     if (
-      current.toHuman().section === 'staking' &&
-      current.toHuman().method === 'nominate'
+      current.method.pallet === 'staking' &&
+      current.method.method === 'nominate'
     ) {
       aggregatedLunieStaking.validators = aggregatedLunieStaking.validators.concat(
-        current.args[0].toHuman()
+        current.args.targets
       )
       hasNominate = true
     }
     reducedMessages.push({
-      section: current.toHuman().section,
-      method: current.toHuman().method,
-      args: JSON.parse(JSON.stringify(current.args, null, 2))
+      section: current.method.pallet,
+      method: current.method.method,
+      args: current.args
     })
   })
   return hasBond && hasNominate
@@ -440,8 +451,8 @@ function coinReducer(network, amount, decimals = 6) {
 function sendDetailsReducer(network, message, signer, reducers) {
   return {
     from: [signer],
-    to: [message.args[0]],
-    amount: reducers.coinReducer(network, message.args[1])
+    to: [message.args.dest],
+    amount: reducers.coinReducer(network, message.args.value)
   }
 }
 
@@ -480,7 +491,7 @@ function extractInvolvedAddresses(
 ) {
   let involvedAddresses = []
   if (lunieTransactionType === lunieMessageTypes.SEND) {
-    involvedAddresses = involvedAddresses.concat([signer, message.args[0]])
+    involvedAddresses = involvedAddresses.concat([signer, message.args.dest])
   } else if (lunieTransactionType === lunieMessageTypes.STAKE) {
     involvedAddresses = involvedAddresses.concat([signer], message.validators)
   } else if (lunieTransactionType === lunieMessageTypes.CLAIM_REWARDS) {
@@ -589,12 +600,9 @@ function networkAccountReducer(address, account, store) {
     }
   }
   return {
-    name:
-      account && account.identity && account.identity.display
-        ? account.identity.display
-        : '',
+    name: account.value ? hexToString(account.value.info.display.Raw) : '',
     address,
-    picture: account ? account.twitter : ''
+    picture: '' // TODO: we need to get images from twitter account
   }
 }
 
@@ -774,8 +782,7 @@ function topVoterReducer(
   network
 ) {
   const councilMemberInfo = electionInfo.members.find(
-    (electionInfoMember) =>
-      electionInfoMember[0].toHuman() === topVoterAddress.toHuman()
+    (electionInfoMember) => electionInfoMember[0].toHuman() === topVoterAddress
   )
   return {
     name: accountInfo.name,
@@ -817,7 +824,6 @@ module.exports = {
   dbRewardsReducer,
   depositReducer,
   networkAccountReducer,
-  getExtrinsicSuccess,
   identityReducer,
   democracyProposalReducer,
   democracyReferendumReducer,